Fixed problem with unsubscribing inside of event notification
Bug was that mission to destroy Dwarg station was not getting recorded
correctly. The problem was that the Dwarg station had multiple
subscribers (zoanthropes also subscribe) and in the handler, the code
removed a subscription.

This shifted the subscriber array, which meant that some notifications
were not sent out.

The fix was to make a copy of the subscription array before dispatching.

Fixed all instances of notifications on subscriber array.