Show quantity of item to use on armor, make Done default if unable to use.
Switching between count on the use screen and no prefix on the result screen is a little weird, but I don't have a better idea at the moment. We want it to show on the result screen even if the player doesn't have any more.

I commented out the call to rpgInitScreenSet because it currently does nothing except add the Done action, and cause an error if it already exists. If it's actually needed in the future, maybe we should fix it so it doesn't cause an error in that case.