Autons remember shield level, continue to regen while in item form.
Should be compatible with old saves. Pre-existing auton items will deploy with full shields.

I used an OnUpdate event, but it only modifies them when they reach full shields so they can coalesce, and only if no dockscreen is open, to avoid bugs when trying to modify them with the auton bay and general confusion. So it's not vital to the functionality and could be removed if it causes performance impact.