Improve placement of primary Charon strongholds in binary systems
It was based on the distance from the center of the system to farthest planet. I changed it to the distance between the closest star to the Korolov station and the farthest planet from that star that's closer to that star than any other star. I also made it avoid being along the line between stars.

https://ministry.kronosaur.com/record.hexm?id=69811